Finished main layout. Just missing to implement background gemotries
