+++ test/app/pages/login/login_page_test.dart
@@ -1,8 +1,4 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:bloc_pattern/bloc_pattern_test.dart';
-
-import 'package:snowmanlabs/app/pages/login/login_page.dart';
 
 main() {
   testWidgets('LoginPage has title', (WidgetTester tester) async {
